+++ astroquery/imcce/tests/test_miriade.py
@@ -42,10 +42,11 @@ def patch_request(request):
 
 def test_spherical_coordinates(patch_request):
     eph = Miriade.get_ephemerides('3552', coordtype=1)
-    cols = ('target', 'epoch', 'RA', 'DEC', 'delta', 'V', 'alpha', 'elong',
-            'RAcosD_rate', 'DEC_rate', 'delta_rate')
-    units = (None, u.d, u.deg, u.deg, u.au, u.mag, u.deg, u.deg,
-             u.arcsec / u.minute, u.arcsec / u.minute, u.km / u.s)
+    cols = ('epoch', 'RA', 'DEC', 'delta', 'V', 'alpha', 'elong', 'RAcosD_rate',
+            'DEC_rate', 'delta_rate', 'rvc', 'berv', 'rvs')
+    units = (u.d, u.deg, u.deg, u.au, u.mag, u.deg, u.deg,
+             u.arcsec / u.minute, u.arcsec / u.minute, u.km / u.s,
+             u.km / u.s, u.km / u.s, u.km / u.s)
     for i in range(len(cols)):
         assert cols[i] in eph.columns
         assert eph[cols[i]].unit == units[i]
@@ -53,14 +54,11 @@ def test_spherical_coordinates(patch_request):
 
 def test_rectangular_coordinates(patch_request):
     eph = Miriade.get_ephemerides('3552', coordtype=2)
-    cols = ('target', 'epoch', 'x', 'y', 'z',
-            'vx', 'vy', 'vz', 'delta', 'V',
-            'alpha', 'elong', 'rv', 'heldist',
-            'x_h', 'y_h', 'z_h',
-            'vx_h', 'vy_h', 'vz_h')
-    units = (None, u.d, u.au, u.au, u.au, u.au/u.day, u.au/u.day,
-             u.au/u.day, u.au, u.mag, u.deg, u.deg, u.km/u.s,
-             u.au, u.au, u.au, u.au, u.au/u.day, u.au/u.day, u.au/u.day)
+    cols = ('epoch', 'px', 'py', 'pz', 'delta', 'heldist', 'alpha', 'elong', 'V',
+            'vx', 'vy', 'vz', 'delta_rate', 'rvc', 'berv', 'rvs')
+    units = (u.d, u.au, u.au, u.au, u.au, u.au, u.deg, u.deg, u.mag,
+             u.au / u.day, u.au / u.day, u.au / u.day,
+             u.km / u.s, u.km / u.s, u.km / u.s, u.km / u.s)
     for i in range(len(cols)):
         assert cols[i] in eph.columns
         assert eph[cols[i]].unit == units[i]
@@ -68,8 +66,11 @@ def test_rectangular_coordinates(patch_request):
 
 def test_local_coordinates(patch_request):
     eph = Miriade.get_ephemerides('3552', coordtype=3)
-    cols = ('target', 'epoch', 'AZ', 'EL', 'V', 'delta', 'alpha', 'elong')
-    units = (None, u.day, u.deg, u.deg, u.mag, u.au, u.deg, u.deg)
+    cols = ('epoch', 'siderealtime', 'AZ', 'EL', 'delta', 'V',
+            'alpha', 'elong', 'refrac', 'delta_rate', 'rvc', 'berv', 'rvs')
+    units = (u.d, 'h min s', u.deg, u.deg, u.au, u.mag,
+             u.deg, u.deg, u.arcsec, u.km / u.s, u.km / u.s,
+             u.km / u.s, u.km / u.s)
     for i in range(len(cols)):
         assert cols[i] in eph.columns
         assert eph[cols[i]].unit == units[i]
@@ -77,9 +78,11 @@ def test_local_coordinates(patch_request):
 
 def test_hourangle_coordinates(patch_request):
     eph = Miriade.get_ephemerides('3552', coordtype=4)
-    cols = ('target', 'epoch', 'hourangle',
-            'DEC', 'V', 'delta', 'alpha', 'elong')
-    units = (None, u.d, u.deg, u.deg, u.mag, u.au, u.deg, u.deg)
+    cols = ['epoch', 'siderealtime', 'hourangle', 'dec', 'delta', 'V',
+            'alpha', 'elong', 'airmass', 'delta_rate', 'rvc', 'berv', 'rvs']
+    units = (u.d, 'h min s', u.deg, u.deg, u.au, u.mag,
+             u.deg, u.deg, None, u.km / u.s, u.km / u.s,
+             u.km / u.s, u.km / u.s)
     for i in range(len(cols)):
         assert cols[i] in eph.columns
         assert eph[cols[i]].unit == units[i]
@@ -87,26 +90,13 @@ def test_hourangle_coordinates(patch_request):
 
 def test_observation_coordinates(patch_request):
     eph = Miriade.get_ephemerides('3552', coordtype=5)
-    cols = ('target', 'epoch', 'siderealtime', 'RAJ2000', 'DECJ2000',
-            'hourangle', 'DEC', 'AZ', 'EL', 'refraction',
-            'V', 'delta', 'heldist', 'alpha', 'elong', 'posunc',
-            'RAcosD_rate', 'DEC_rate', 'delta_rate')
-    units = (None, u.d, u.h, u.deg, u.deg, u.deg, u.deg, u.deg, u.deg,
-             u.arcsec, u.mag, u.au, u.au, u.deg, u.deg, u.arcsec,
-             u.arcsec / u.minute, u.arcsec / u.minute, u.km / u.s)
-    for i in range(len(cols)):
-        assert cols[i] in eph.columns
-        assert eph[cols[i]].unit == units[i]
-
-
-def test_aoobservation_coordinates(patch_request):
-    eph = Miriade.get_ephemerides('3552', coordtype=6)
-    cols = ('target', 'epoch', 'siderealtime', 'RAJ2000', 'DECJ2000',
-            'refraction', 'V', 'delta', 'heldist', 'alpha',
-            'elong', 'posunc', 'RAcosD_rate', 'DEC_rate', 'delta_rate')
-    units = (None, u.d, u.h, u.deg, u.deg, u.arcsec, u.mag,
-             u.au, u.au, u.deg, u.deg, u.arcsec, u.arcsec / u.minute,
-             u.arcsec / u.minute, u.km / u.s)
+    cols = ['epoch', 'siderealtime', 'RA', 'DEC', 'hourangle', 'AZ', 'EL',
+            'delta', 'heldist', 'V', 'alpha', 'elong', 'airmass',
+            'RAcosD_rate', 'DEC_rate', 'delta_rate', 'rvc', 'berv', 'rvs']
+    units = (u.d, "h min s", u.deg, u.deg, u.deg, u.deg, u.deg,
+             u.au, u.au, u.mag, u.deg, u.deg, None,
+             u.arcsec / u.min, u.arcsec / u.min, u.km / u.s, u.km / u.s,
+             u.km / u.s, u.km / u.s)
     for i in range(len(cols)):
         assert cols[i] in eph.columns
         assert eph[cols[i]].unit == units[i]
@@ -115,4 +105,4 @@ def test_aoobservation_coordinates(patch_request):
 def test_get_raw_response(patch_request):
     raw_eph = Miriade.get_ephemerides(
         '3552', coordtype=1, get_raw_response=True)
-    assert "<?xml version='1.0' encoding='UTF-8'?>" in raw_eph
+    assert '<?xml version="1.0" encoding="UTF-8"?>' in raw_eph