+++ CHANGES.rst
@@ -64,6 +64,13 @@ esa.hubble
 - Update ``get_datalabs_path`` method so an alternative path is checked if the
   file is not in Datalabs yet [#3437]
 
+imcce
+^^^^^
+
+- Change the URL for SkyBot and Miriade Web Services
+- Adapted the ``Miriade`` Class to the new outputs of the Web Service
+
+
 mast
 ^^^^